Split modal manifest into core + (app), enable verbatimModuleSyntax
Two related changes:

1) verbatimModuleSyntax: true

Turn on the TS flag that forces every type-only import to use
`import type` syntax. Zero errors trigger across the monorepo —
the codebase already follows this convention — but enabling the
flag locks the discipline in so future regressions get caught at
typecheck time. Useful guardrail against accidentally pulling
runtime code through what was meant to be a type-only path.

2) Modal manifest split

Previously `lib/modal/modals/index.ts` registered every modal
(27 total) in a single `availableModals` object imported by
GlobalModalProvider in the root layout. Even with each modal
wrapped in next/dynamic(), Next.js still bundled the modals' JS
chunks into routes that never render them (blog, articles,
landing) because the manifest was statically loaded everywhere.

Replace with a runtime registry split into:

- modals/core.ts — 7 always-loaded modals (confirmation, info,
  auth/connection/session errors, rate-limit, example). Seeded
  into the registry at module load.
- modals/app.ts — 21 (app)-only modals (exercise, premium,
  subscription, payment, settings, badge, welcome, walkthrough).
- modals/registry.ts — module-level registry with registerModals()
  and getModal().
- AppModalRegistrar.tsx — client component that calls
  registerModals(appModals) as a side effect at module load.

(app)/layout.tsx mounts <AppModalRegistrar />, so the (app)
bundle picks up the app modals on hydration. Non-(app) routes
never import modals/app.ts and never pay for the 21 modal chunks.

GlobalModalProvider now reads from the registry via getModal()
instead of the static manifest.

Known follow-up: lib/modal/store.ts still imports CSS modules
from several (app)-only modals at the top level (PremiumUpgrade,
SubscriptionCheckout, Payment*, WelcomeToPremium, Welcome,
WalkthroughConfirm, AvatarEdit). Because the store is loaded
globally via useModalStore, those CSS modules continue to bundle
into the global CSS chunk. The JS split lands here; decoupling
store.ts from modal CSS is a separate refactor.
