+++ app/app/(app)/layout.tsx
@@ -1,6 +1,7 @@
 import { ClientAuthGuard } from "../../components/layout/auth/internal/ClientAuthGuard";
 import { CheckoutReturnHandler } from "@/components/checkout/CheckoutReturnHandler";
 import { WelcomeModalHandler } from "@/components/WelcomeModalHandler";
+import { AppModalRegistrar } from "@/lib/modal/AppModalRegistrar";
 
 /**
  * Internal App Layout with Client-Side Authentication
@@ -32,6 +33,7 @@ export default function AppLayout({
 }>) {
   return (
     <ClientAuthGuard>
+      <AppModalRegistrar />
       {children}
       <CheckoutReturnHandler />
       <WelcomeModalHandler />
