+++ app/lib/modal/modals/app.ts
@@ -1,17 +1,16 @@
 import dynamic from "next/dynamic";
 
-const ConfirmationModal = dynamic(() => import("./ConfirmationModal").then((m) => m.ConfirmationModal));
-const ConnectionErrorModal = dynamic(() => import("./ConnectionErrorModal").then((m) => m.ConnectionErrorModal));
-const AuthErrorModal = dynamic(() => import("./AuthErrorModal").then((m) => m.AuthErrorModal));
-const ExampleModal = dynamic(() => import("./ExampleModal").then((m) => m.ExampleModal));
+// (app)-only modals: triggered exclusively from authenticated routes
+// (exercise flows, premium upgrade, subscription/payment, settings,
+// badges, welcome). Registered by (app)/layout.tsx via AppModalRegistrar
+// so the corresponding JS chunks aren't shipped to /blog, /articles,
+// /, /premium etc.
+
 const ExerciseSuccessModal = dynamic(() => import("./ExerciseSuccessModal").then((m) => m.ExerciseSuccessModal));
 const ExerciseCompletionModal = dynamic(() =>
   import("./ExerciseCompletionModal").then((m) => m.ExerciseCompletionModal)
 );
 const LevelMilestoneModal = dynamic(() => import("./LevelMilestoneModal").then((m) => m.LevelMilestoneModal));
-const InfoModal = dynamic(() => import("./InfoModal").then((m) => m.InfoModal));
-const SessionExpiredModal = dynamic(() => import("./SessionExpiredModal").then((m) => m.SessionExpiredModal));
-const RateLimitModal = dynamic(() => import("./RateLimitModal").then((m) => m.RateLimitModal));
 const SubscriptionModal = dynamic(() => import("./SubscriptionModal").then((m) => m.SubscriptionModal));
 const SubscriptionSuccessModal = dynamic(() =>
   import("./SubscriptionSuccessModal").then((m) => m.SubscriptionSuccessModal)
@@ -51,18 +50,10 @@ const WalkthroughConfirmModal = dynamic(() =>
 );
 const WelcomeModal = dynamic(() => import("./WelcomeModal").then((m) => m.WelcomeModal));
 
-// Available modals registry
-export const availableModals = {
-  "example-modal": ExampleModal,
-  "confirmation-modal": ConfirmationModal,
-  "connection-error-modal": ConnectionErrorModal,
-  "auth-error-modal": AuthErrorModal,
-  "info-modal": InfoModal,
+export const appModals = {
   "exercise-success-modal": ExerciseSuccessModal,
   "exercise-completion-modal": ExerciseCompletionModal,
   "level-milestone-modal": LevelMilestoneModal,
-  "session-expired-modal": SessionExpiredModal,
-  "rate-limit-modal": RateLimitModal,
   "subscription-modal": SubscriptionModal,
   "subscription-success-modal": SubscriptionSuccessModal,
   "subscription-checkout-modal": SubscriptionCheckoutModal,
